Implement Hopfield network for question 3.1
